pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xadl"
version = "0.1.0"
description = "Explainable Contrastive Anomaly Detection"
authors = [{name="Your Name"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch>=2.2",
  "numpy>=1.26",
  "pandas>=2.0",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
  "tqdm>=4.66"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
